New Changes Java to DB
Created new java clas DBUpdater. Contains a method that will increment number of items in report table given the primary key as parameters: (BotID, ItemID).

Implemented DBUpdater into "MyFirstScript" which is a bot script. Everytime bot cuts a tree that update method in DBUpdater is called. Must run main() simultaneously for this to work since run() did not like DBUpdater. Instead, bot code will writer to a file simultaneous to when it is reading it. Bot code will write "1" if wood was collected, "0" other wise. Main() reads this file and will call on the update method if it reads "1", else nothing happens.